Update Java Code
This is a coding segment of a previous java project. 
The background of the project was This project focuses on managing travel between planets at the Wayward Planets Resort, which is a part of the WISH organization that offers holiday packages in space. The Wayward Planets Resort consists of multiple planets, each with different leisure facilities, and guests are provided with personalised permits loaded with credits to travel between the planets via shuttles.
The project implemented a system that ensured compliance with the conditions for using permits, enabled guests to purchase credits at any time and managed travel between the planets efficiently.