Cisco Packet Tracer
This project, which involved setting up a artificial network for a fictional small network at the Hatfield branch, I was tasked with implementing and configuring a computer network for the Hatfield branch of TA Finance, a finance company that provides loans to individuals and businesses. The company is expanding its small branches in the high street, and depending on the physical layout of each office, the network team designs a computer network to fulfil its needs. The network topology for the Hatfield branch had a router, a internet service, cloud storage, access points, switches and wired end devices and is designed to support IPv4 and ensure full connectivity to the headquarters via the internet.