Netlogo Scala code
In this example code the bugs in the code have to pick up the chip if it lands on a green pixel and put it next to the biggest object or circle. This can be configured and can have 2 different circles and where the circles can be placed. Also there is a function where you can draw where each bug goes and track how they move.